# feat: support `~/.plutorc` for options