This PR greatly simplifies the duplication in `Pluto.Configuration.from_flat_kwargs` and adds a layer of reading options from a file, in a seamless way.
## Try this Pull Request!
Open Julia and type:
```jl
julia> import Pkg
julia> Pkg.activate(temp=true)
julia> Pkg.add(url="https://github.com/fonsp/Pluto.jl", rev="pg/pluto-persistent-options")
julia> using Pluto
```