@fonsp 2025-04-25
Hey @pankgeorg thanks for looking into this! Unfortunately this is some pretty rigid code...

For the duplication: this was done to fix a performance issue: https://github.com/fonsp/Pluto.jl/pull/2005 If we remove it, we also need to check these benchmarks again. (Might have changed for new Julia versions.) For me it's not worth the time, as LLMs can work with this file really well and we don't need to change it often.

For `~/.plutorc`: the two issues here are:
- We already have a config file format (for PlutoSliderServer) that can configure Pluto: PlutoDeployment.toml, so I want to avoid adding a new one.
- After we implemented Configurations.jl and PlutoDeployment.toml in Pluto, Julia came out with Preferences.jl and LocalPreferences.toml. 

So if we change something to Pluto's config system, it would be moving to Preferences.jl and LocalPreferences.toml (maybe Configurations.jl can help with this) and supporting Julia 1.12 module arguments (https://github.com/Roger-luo/Configurations.jl/issues/100). This would probably be a breaking change, one that will take a lot of time across the Pluto ecosystem, and I'm not sure if it is worth the trade-off.