@pankgeorg 2025-04-25
Performance wise it's not good enough at this point

main:
```
julia> import Pluto
Precompiling Pluto...
  1 dependency successfully precompiled in 26 seconds. 68 already precompiled.

julia> @time @eval Pluto.Configuration.from_flat_kwargs();
  0.000244 seconds (82 allocations: 3.250 KiB)
```

this PR: 
```
julia> import Pluto
Precompiling Pluto...
  1 dependency successfully precompiled in 29 seconds. 68 already precompiled.

julia> @time @eval Pluto.Configuration.from_flat_kwargs();
[ Info: Using options from ~/.plutorc
  1.228102 seconds (2.13 M allocations: 107.029 MiB, 13.95% gc time, 99.74% compilation time)

julia> @time @eval Pluto.Configuration.from_flat_kwargs();
[ Info: Using options from ~/.plutorc
  0.000946 seconds (664 allocations: 36.703 KiB)
```

I'll report improvements over this soon!